Refactor report_engine.py into modular builders and add frontend tests for report-result-card.tsx
