+++ apps/web/jest.config.ts
@@ -1,0 +1,17 @@
+import type { Config } from 'jest';
+import nextJest from 'next/jest.js';
+
+const createJestConfig = nextJest({
+    dir: './',
+});
+
+const config: Config = {
+    coverageProvider: 'v8',
+    testEnvironment: 'jsdom',
+    setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
+    moduleNameMapper: {
+        '^@/(.*)$': '<rootDir>/src/$1',
+    },
+};
+
+export default createJestConfig(config);
